refactor(useScrollAnimation): hook으로 분리
